refactor(racingcar): do/while문을 재귀문으로 변경
